Widen scopes_per_resource and metrics_per_scope to u16
The previous u8 typing capped the per-emission cardinality at 255×255 =
65,025 unique metric sites per resource. Widening these two fields to
u16 raises the ceiling to ~4.3B, which fits within the existing
total_contexts: u32 envelope. Validation arithmetic uses
saturating_mul to avoid u32 overflow at the upper bound.

+++ lading_payload/src/opentelemetry/metric/templates.rs
@@ -285,7 +285,7 @@ pub(crate) enum Kind {
 
 #[derive(Clone, Debug)]
 pub(crate) struct ScopeTemplateGenerator {
-    metrics_per_scope: ConfRange<u8>,
+    metrics_per_scope: ConfRange<u16>,
     metric_generator: MetricTemplateGenerator,
     str_pool: Rc<strings::RandomStringPool>,
     tags: TagGenerator,
@@ -404,7 +404,7 @@ impl<'a> crate::SizedGenerator<'a> for ScopeTemplateGenerator {
 
 #[derive(Clone, Debug)]
 pub(crate) struct ResourceTemplateGenerator {
-    scopes_per_resource: ConfRange<u8>,
+    scopes_per_resource: ConfRange<u16>,
     attributes_per_resource: ConfRange<u8>,
     scope_generator: ScopeTemplateGenerator,
     tags: TagGenerator,
